requirements.txt
Django==1.8
djangorestframework==3.5.1
django-cors-headers==2.1.0
django-rest-swagger==2.1.0
